==> fine-tuning-embeddings/4-use_finetuned_embeddings.ipynb <==
# Use Fine-tuned Model

Based on example from [here](https://sbert.net/docs/sentence_transformer/training_overview.html#trainer)

### imports

from sentence_transformers import SentenceTransformer
from datasets import load_dataset
import numpy as np

### load model

model = SentenceTransformer("pnimeesha/distilroberta-ai-job-embeddings")

### load data

dataset = load_dataset("pnimeesha/ai-job-embedding-finetuning")

### use it!

# new query
query = "data scienist 6 year experience, LLMs, credit risk, content marketing"
query_embedding = model.encode(query)

# encode JDs
jd_embeddings = model.encode(dataset["test"]["job_description_pos"])

# compute similarities
similarities = model.similarity(query_embedding, jd_embeddings)
print(similarities.shape)

# sort JDs based on similarity
simiarities_argsorted = np.argsort(similarities.numpy(), axis=1)

# return most similar job description
dataset["test"]["job_description_pos"][int(simiarities_argsorted[0][-1])]
